requirements.txt
backports.functools-lru-cache==1.6.4
bio==1.5.9
black==22.6.0
bottleneck==1.3.5
ete3==3.1.2
importlib-metadata==6.1.0
ipykernel==6.15.0
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.4
phylopandas==0.8.0
pip-chill==1.0.1
pydantic==1.10.7
pysqlite3==0.5.0
sqlalchemy==1.4.39
